+++ R/coroutine.R
@@ -1,33 +1,3 @@
-# Some generics and methods common to all coroutines.
-
-#' @rdname format
-#' @param x a coroutine ([async], [gen], or [stream]) object.
-#' @return `getNode()` returns a character naming the last known execution node.
-#' @param ... Undocumented.
-#' @details
-#' The state displayed by `format()` or `getNode()` can be read like an
-#' adsress pointing to a spot in the source code; for example, a state
-#' string like `.{1.<-2.await__then` can be read like
-#' "in the first argument of `{`, in the second argument of `<-`, in a
-#' call to `await()`, at internal node `then`."
-#'
-#' @export
-getNode <- function(x, ...) UseMethod("getNode")
-
-#' @export
-#' @rdname format
-getState <- function(x, ...) UseMethod("getState")
-
-#' @export
-#' @rdname format
-#' @return `getOrig()` returns the original expression given to the coroutine's
-#' constructor.
-getOrig <- function(x, ...) UseMethod("getOrig")
-
-#' @export
-#' @rdname format
-#' @return `getEnv()` returns the coroutine's effective environment.
-getEnv <- function(x, ...) UseMethod("getEnv")
 
 # not-exported methods
 getPump <- function(x) UseMethod("getPump")
@@ -40,11 +10,6 @@ compile <- function(x, level, ...) UseMethod("compile")
 reconstitute <- function(orig, munged) UseMethod("reconstitute")
 
 #' @exportS3Method
-getEnv.coroutine <- function(x, ...) {
-  environment(getPump(x))
-}
-
-#' @exportS3Method
 getReturn.coroutine <- function(x) {
   environment(getPump(x))$return_
 }
@@ -52,12 +17,6 @@ getReturn.coroutine <- function(x) {
 #' @exportS3Method
 getStop.coroutine <- function(x) {
   environment(getPump(x))$stop_
-}
-
-#' @exportS3Method
-#' @rdname format
-getNode.coroutine <- function(x, ...) {
-  environment(getPump(x))$getCont()
 }
 
 #' @exportS3Method
@@ -71,10 +30,14 @@ print.coroutine <- function(x, ...) {
 }
 
 #' Toggle single-step debugging for a coroutine.
-#' @param x A coroutine object as constructed by ([async], [gen] or [stream]).
-#' @param R set TRUE to step through expressions at user level
-#' @param internal Set TRUE to step through at coroutine implementation level.
-#' @return a `list(R=, internal=) with the current debug setting.
+#' @param x A coroutine object as constructed by ([async], [gen] or
+#'   [stream]).
+#' @param R Set TRUE to step through expressions at user level
+#' @param internal Set TRUE to step through at coroutine
+#'   implementation level.
+#' @param trace Set TRUE or provide a print function to print each R
+#'   expression evaluated in turn.
+#' @return a `list(R=, internal=, trace=)` with the current debug state.
 #' @export
 debugAsync <- function(x, R, internal, trace) UseMethod("debugAsync")
 
@@ -86,7 +49,7 @@ debugAsync.coroutine <- function(x, R=current$R, internal=current$internal, trac
 }
 
 #' @exportS3Method
-compile.coroutine <- function(x, level) {
+compile.coroutine <- function(x, level, ...) {
    if (abs(level) >= 1) {
     if (getOption("async.paranoid")) graph <- walk(x)
     munged <- munge( x )
@@ -122,26 +85,22 @@ getStartSet.coroutine <- function(x) {
 
 #' Query / display coroutine properties and state.
 #'
-#' A coroutine's `format` method displays its pre-compiled source
-#'   code, its effective environment, whether it is running or
-#'   finished, and a label indicating its last known state. The
-#'   methods `getState`, `getNode` and `getOrig` also expose this
-#'   information.
-#'
+#' The coroutine `format` method displays its source code, its
+#'   effective environment, whether it is running or finished, and a
+#'   label indicating its last known state. The `summary` method
+#'   returns the same information in a list.
+#' @param x A coroutine.
 #' @exportS3Method
 #' @rdname format
 format.coroutine <- function(x, ...) {
-  envir <- getEnv(x)
-  code <- getOrig(x)
-  a <- deparse(call(class(x)[[1]], code), backtick=TRUE)
-  b <- format(envir, ...)
-  state <- getState(x)
-  cont <- getNode(x)
+  s <- summary(x)
+  a <- deparse(call(class(x)[[1]], s$code), backtick=TRUE)
+  b <- format(s$envir, ...)
   c <- paste0(c("<", class(x)[[1]], " [",
-                state,
-                " at `", cont, "`",
-                if (state=="stopped")
-                  c(": ", capture.output(print(envir$err))),
+                s$state,
+                " at `", s$node, "`",
+                if (s$state=="stopped")
+                  c(": ", capture.output(print(s$err))),
                 "]>"), collapse="")
   d <- NextMethod()
   c(a, b, c, d)
@@ -160,4 +119,29 @@ coroutine_function <- function(arg, head, ...) {
                        as.call(c(list(head, body, local=FALSE), extra))))
   fn <- call("function", args, body)
   quo_(fn, env(arg))
+}
+
+#' @rdname format
+#' @param object a coroutine ([async], [generator][gen], or [stream]) object.
+#' @param ... Undocumented.
+#' @description `summary(obj)` returns a list with information on a coroutine's state,  including:
+#' * `code`: the expression used to create the coroutine;
+#' * `state`: the current state (see below);
+#' * `node`: is a character string that
+#'    identifies a location in the coroutine source code; for example,
+#'    a typical state string might be ".\{.<-2.await__then", which
+#'    can be read like "in the first argument of `\{`, in the second
+#'    argument of `<-`, in a call to `await()`, at internal node `then`.";
+#' * `envir`: the environment where the coroutine is evaluating R expressions;
+#' * `err`: the error object, if the coroutine caught an error.
+#' @export
+summary.coroutine <- function(object, ...) {
+  d <- debugAsync(object)
+  list(
+    node=environment(getPump(object))$getCont(),
+    envir=environment(getPump(object))$targetEnv,
+    err=environment(getPump(object))$err,
+    debugR=d$R,
+    debugInternal=d$internal,
+    trace=d$trace)
 }